refactor to have only top level class in psedata
