+++ src/lib.rs
@@ -1,3 +1,2 @@
-pub mod pymol_structs;
-
-pub use pymol_structs::PSEData;
+pub mod psedata;
+mod pymolparsing;
